+++ ui/src/components/file_transfer/style.scss
